Modulo de Gestion

+++ app/Http/Controllers/UserController.php
@@ -9,37 +9,27 @@ use App\Http\Requests\UserRequest;
 
 class UserController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
+
+    public function index(Request $request)
     {
-        $users = User::orderBy('id', 'ASC')->paginate(10);
+        $users = User::Search($request->name)
+            ->orderBy('id', 'ASC')
+            ->paginate(5);
+
         return view('admin.users.index')->with('users', $users);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function create()
     {
         return view('admin.users.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+
     public function store(Request $request)
     {
         $user = new User($request->all());
-        $user->password = bcrypt($request->password);
+        // $user->password = bcrypt($request->password);
         $user->save();
 
         Flash::success("Se ha registrado " . $user->name . " de forma exitosa!");
@@ -47,36 +37,20 @@ class UserController extends Controller
         return redirect()->route('users.index');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function show($id)
     {
         //
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function edit($id)
     {
         $user = User::find($id);
         return view('admin.users.edit')->with('user', $user);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function update(Request $request, $id)
     {
         $user = User::find($id);
@@ -88,19 +62,17 @@ class UserController extends Controller
         return redirect()->route('users.index');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function destroy($id)
     {
         $user = User::find($id);
-        $user->delete();
 
-        // Flash::error('El usuario ' . $user->name . ' ha sido borrado de forma Exitosa!!');
-
+        if (!count($user->customers)) {
+            $user->delete();
+            Flash::error('El usuario ' . $user->name . ' ha sido borrado de forma Exitosa!!');
+            return redirect()->route('users.index');
+        }
+        Flash::error('El usuario ' . $user->name . ' no se pudo eliminar "Tiene Clientes Asociados"!!');
         return redirect()->route('users.index');
     }
 }
